+++ common/fe-ste-models.go
@@ -23,6 +23,7 @@ package common
 import (
 	"bytes"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"math"
 	"os"
@@ -55,6 +56,8 @@ const (
 	// Since we haven't updated the Go SDKs to handle CPK just yet, we need to detect CPK related errors
 	// and inform the user that we don't support CPK yet.
 	CPK_ERROR_SERVICE_CODE = "BlobUsesCustomerSpecifiedEncryption"
+	BLOB_NOT_FOUND         = "BlobNotFound"
+	FILE_NOT_FOUND         = "The specified file was not found."
 )
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -250,6 +253,7 @@ func (OverwriteOption) True() OverwriteOption          { return OverwriteOption(
 func (OverwriteOption) False() OverwriteOption         { return OverwriteOption(1) }
 func (OverwriteOption) Prompt() OverwriteOption        { return OverwriteOption(2) }
 func (OverwriteOption) IfSourceNewer() OverwriteOption { return OverwriteOption(3) }
+func (OverwriteOption) PosixProperties() OverwriteOption {return OverwriteOption(4)}
 
 func (o *OverwriteOption) Parse(s string) error {
 	val, err := enum.Parse(reflect.TypeOf(o), s, true)
@@ -670,6 +674,11 @@ var ETransferStatus = TransferStatus(0)
 
 type TransferStatus int32 // Must be 32-bit for atomic operations; negative #s represent a specific failure code
 
+func (t TransferStatus) StatusLocked() bool { // Is an overwrite necessary to change tx status?
+	// Any kind of failure, or success is considered "locked in".
+	return t <= ETransferStatus.Failed() || t == ETransferStatus.Success()
+}
+
 // Transfer is ready to transfer and not started transferring yet
 func (TransferStatus) NotStarted() TransferStatus { return TransferStatus(0) }
 
@@ -1077,13 +1086,53 @@ func UnMarshalToCommonMetadata(metadataString string) (Metadata, error) {
 func StringToMetadata(metadataString string) (Metadata, error) {
 	metadataMap := Metadata{}
 	if len(metadataString) > 0 {
-		for _, keyAndValue := range strings.Split(metadataString, ";") { // key/value pairs are separated by ';'
-			kv := strings.Split(keyAndValue, "=") // key/value are separated by '='
-			// what if '=' not present?
-			if len(kv) != 2 {
-				return metadataMap, fmt.Errorf("invalid metadata string passed")
+		cKey := ""
+		cVal := ""
+		keySet := false
+		ignoreRules := false
+
+		addchar := func(c rune) {
+			if !keySet {
+				cKey += string(c)
+			} else {
+				cVal += string(c)
 			}
-			metadataMap[kv[0]] = kv[1]
+		}
+		for _, c := range metadataString {
+			if ignoreRules {
+				addchar(c)
+				ignoreRules = false
+			} else {
+				switch c {
+				case '=':
+					if keySet {
+						addchar(c)
+					} else {
+						keySet = true
+					}
+
+				case ';':
+					if !keySet {
+						return Metadata{}, errors.New("metadata names must conform to C# naming rules (https://learn.microsoft.com/en-us/rest/api/storageservices/naming-and-referencing-containers--blobs--and-metadata#metadata-names)")
+					}
+
+					metadataMap[cKey] = cVal
+					cKey = ""
+					cVal = ""
+					keySet = false
+					ignoreRules = false
+
+				case '\\':
+					ignoreRules = true // ignore the rules on the next character
+
+				default:
+					addchar(c)
+				}
+			}
+		}
+
+		if cKey != "" {
+			metadataMap[cKey] = cVal
 		}
 	}
 	return metadataMap, nil
@@ -1615,4 +1664,29 @@ func (rpt RehydratePriorityType) ToRehydratePriorityType() azblob.RehydratePrior
 	default:
 		return azblob.RehydratePriorityStandard
 	}
+}
+
+// //////////////////////////////////////////////////////////////////////////////
+type SyncHashType uint8
+
+var ESyncHashType SyncHashType = 0
+
+func (SyncHashType) None() SyncHashType {
+	return 0
+}
+
+func (SyncHashType) MD5() SyncHashType {
+	return 1
+}
+
+func (ht *SyncHashType) Parse(s string) error {
+	val, err := enum.ParseInt(reflect.TypeOf(ht), s, true, true)
+	if err == nil {
+		*ht = val.(SyncHashType)
+	}
+	return err
+}
+
+func (ht SyncHashType) String() string {
+	return enum.StringInt(ht, reflect.TypeOf(ht))
 }